# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FootballLive API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches/{id}/lineup/auto:
    post:
      operationId: autoFillLineup
      tags: [lineup]
      summary: Auto-fill match lineup from both teams' squads
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lineup auto-filled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchLineup"
...
  /matches/{id}/lineup:
    get:
      operationId: getMatchLineup
      tags: [lineup]
      summary: Get match lineup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lineup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchLineup"
    post:
      operationId: addLineupPlayer
      tags: [lineup]
      summary: Add a player to the lineup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LineupPlayerInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineupPlayer"
  /matches/{id}/lineup/{playerId}:
    patch:
      operationId: updateLineupPlayer
      tags: [lineup]
      summary: Update a lineup player's role or isStarting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LineupPlayerPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineupPlayer"
    delete:
      operationId: removeLineupPlayer
      tags: [lineup]
      summary: Remove a player from lineup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /highlights:
    get:
      operationId: listHighlights
      tags: [highlights]
      summary: List all highlights
      parameters:
        - name: competition
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Highlights list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Highlight"
    post:
      operationId: createHighlight
      tags: [highlights]
      summary: Create a highlight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HighlightInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
  /highlights/{id}:
    get:
      operationId: getHighlight
      tags: [highlights]
      summary: Get a single highlight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Highlight found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
    delete:
      operationId: deleteHighlight
      tags: [highlights]
      summary: Delete a highlight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name, shortName, logoUrl, country, sport]
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        logoUrl:
          type: string
        country:
          type: string
        sport:
          type: string
          enum: [football, futsal]
...
    LineupPlayer:
      type: object
      required: [id, matchId, teamId, playerNumber, playerName, role, isStarting]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        teamId:
          type: integer
        playerNumber:
          type: string
        playerName:
          type: string
        position:
          type: ["string", "null"]
        role:
          type: string
        isStarting:
          type: boolean
    LineupPlayerPatch:
      type: object
      properties:
        role:
          type: string
        isStarting:
          type: boolean
    LineupPlayerInput:
      type: object
      required: [teamId, playerNumber, playerName]
      properties:
        teamId:
          type: integer
        playerNumber:
          type: string
        playerName:
          type: string
        position:
          type: string
        role:
          type: string
        isStarting:
          type: boolean
    MatchLineup:
      type: object
      required: [matchId, home, away]
      properties:
        matchId:
          type: integer
        home:
          type: array
          items:
            $ref: "#/components/schemas/LineupPlayer"
        away:
          type: array
          items:
            $ref: "#/components/schemas/LineupPlayer"
...
    Highlight:
      type: object
      required: [id, title, competition, thumbnailUrl, videoUrl, duration, publishedAt, homeTeam, awayTeam, homeScore, awayScore]
      properties:
        id:
          type: integer
        title:
          type: string
        competition:
          type: string
        thumbnailUrl:
          type: string
        videoUrl:
          type: string
        duration:
          type: string
        publishedAt:
          type: string
          format: date-time
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        homeScore:
          type: integer
        awayScore:
          type: integer
        views:
          type: integer
    HighlightInput:
      type: object
      required: [title, competition, thumbnailUrl, videoUrl, duration, homeTeamId, awayTeamId, homeScore, awayScore]
      properties:
        title:
          type: string
        competition:
          type: string
        thumbnailUrl:
          type: string
        videoUrl:
          type: string
        duration:
          type: string
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        homeScore:
          type: integer
        awayScore:
          type: integer
...